Create edit folder and move EditGallery
